Update plat_trophies.cpp
changed the function from void to a string and returned the gameName in the if clause and at the end of the string function itself. Also added a variable named "game" and set it equal to the function call of EnterGame. The outfile now has the "game" variable so the getline function actually works